+++ crates/miden-objects/src/transaction/tx_args.rs
@@ -1,11 +1,11 @@
-use alloc::collections::{BTreeMap, BTreeSet};
+use alloc::collections::BTreeMap;
 use alloc::sync::Arc;
 use alloc::vec::Vec;
 
 use miden_crypto::merkle::InnerNodeInfo;
 use miden_processor::MastNodeExt;
 
-use super::{AccountInputs, Felt, Hasher, Word};
+use super::{Felt, Hasher, Word};
 use crate::account::{PublicKeyCommitment, Signature};
 use crate::note::{NoteId, NoteRecipient};
 use crate::utils::serde::{
@@ -39,13 +39,12 @@ use crate::{EMPTY_WORD, MastForest, MastNodeId};
 ///   this argument is not specified, the [`EMPTY_WORD`] would be used as a default value. If the
 ///   [AdviceInputs] are propagated with some user defined map entries, this argument could be used
 ///   as a key to access the corresponding value.
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq)]
 pub struct TransactionArgs {
     tx_script: Option<TransactionScript>,
     tx_script_args: Word,
     note_args: BTreeMap<NoteId, Word>,
     advice_inputs: AdviceInputs,
-    foreign_account_inputs: Vec<AccountInputs>,
     auth_args: Word,
 }
 
@@ -55,7 +54,7 @@ impl TransactionArgs {
 
     /// Returns new [TransactionArgs] instantiated with the provided transaction script, advice
     /// map and foreign account inputs.
-    pub fn new(advice_map: AdviceMap, foreign_account_inputs: Vec<AccountInputs>) -> Self {
+    pub fn new(advice_map: AdviceMap) -> Self {
         let advice_inputs = AdviceInputs { map: advice_map, ..Default::default() };
 
         Self {
@@ -63,7 +62,6 @@ impl TransactionArgs {
             tx_script_args: EMPTY_WORD,
             note_args: Default::default(),
             advice_inputs,
-            foreign_account_inputs,
             auth_args: EMPTY_WORD,
         }
     }
@@ -138,19 +136,6 @@ impl TransactionArgs {
     /// Returns a reference to the internal [AdviceInputs].
     pub fn advice_inputs(&self) -> &AdviceInputs {
         &self.advice_inputs
-    }
-
-    /// Returns a reference to the foreign account inputs in the transaction arguments.
-    pub fn foreign_account_inputs(&self) -> &[AccountInputs] {
-        &self.foreign_account_inputs
-    }
-
-    /// Collects and returns a set containing all code commitments from foreign accounts.
-    pub fn to_foreign_account_code_commitments(&self) -> BTreeSet<Word> {
-        self.foreign_account_inputs()
-            .iter()
-            .map(|acc| acc.code().commitment())
-            .collect()
     }
 
     /// Returns a reference to the authentication procedure argument, or [`EMPTY_WORD`] if the
@@ -239,13 +224,18 @@ impl TransactionArgs {
     }
 }
 
+impl Default for TransactionArgs {
+    fn default() -> Self {
+        Self::new(AdviceMap::default())
+    }
+}
+
 impl Serializable for TransactionArgs {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.tx_script.write_into(target);
         self.tx_script_args.write_into(target);
         self.note_args.write_into(target);
         self.advice_inputs.write_into(target);
-        self.foreign_account_inputs.write_into(target);
         self.auth_args.write_into(target);
     }
 }
@@ -256,7 +246,6 @@ impl Deserializable for TransactionArgs {
         let tx_script_args = Word::read_from(source)?;
         let note_args = BTreeMap::<NoteId, Word>::read_from(source)?;
         let advice_inputs = AdviceInputs::read_from(source)?;
-        let foreign_account_inputs = Vec::<AccountInputs>::read_from(source)?;
         let auth_args = Word::read_from(source)?;
 
         Ok(Self {
@@ -264,7 +253,6 @@ impl Deserializable for TransactionArgs {
             tx_script_args,
             note_args,
             advice_inputs,
-            foreign_account_inputs,
             auth_args,
         })
     }
@@ -347,7 +335,7 @@ mod tests {
 
     #[test]
     fn test_tx_args_serialization() {
-        let tx_args = TransactionArgs::new(AdviceMap::default(), std::vec::Vec::default());
+        let tx_args = TransactionArgs::new(AdviceMap::default());
         let bytes: std::vec::Vec<u8> = tx_args.to_bytes();
         let decoded = TransactionArgs::read_from_bytes(&bytes).unwrap();