+++ crates/miden-testing/src/kernel_tests/tx/test_epilogue.rs
@@ -9,33 +9,25 @@ use miden_lib::errors::tx_kernel_errors::{
 };
 use miden_lib::testing::mock_account::MockAccountExt;
 use miden_lib::testing::note::NoteBuilder;
+use miden_lib::transaction::EXPIRATION_BLOCK_ELEMENT_IDX;
 use miden_lib::transaction::memory::{
     NOTE_MEM_SIZE,
     OUTPUT_NOTE_ASSET_COMMITMENT_OFFSET,
     OUTPUT_NOTE_SECTION_OFFSET,
 };
-use miden_lib::transaction::{EXPIRATION_BLOCK_ELEMENT_IDX, TransactionKernel};
 use miden_lib::utils::ScriptBuilder;
 use miden_objects::Word;
 use miden_objects::account::{Account, AccountDelta, AccountStorageDelta, AccountVaultDelta};
-use miden_objects::asset::{Asset, AssetVault, FungibleAsset};
+use miden_objects::asset::{Asset, FungibleAsset};
 use miden_objects::note::{NoteTag, NoteType};
 use miden_objects::testing::account_id::{
     ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1,
-    ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_2,
-    ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_3,
     ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE,
     ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE,
     ACCOUNT_ID_SENDER,
 };
-use miden_objects::testing::constants::{
-    CONSUMED_ASSET_1_AMOUNT,
-    CONSUMED_ASSET_2_AMOUNT,
-    CONSUMED_ASSET_3_AMOUNT,
-};
 use miden_objects::transaction::{OutputNote, OutputNotes};
 use miden_processor::{Felt, ONE};
-use rand::rng;
 
 use super::{ZERO, create_mock_notes_procedure};
 use crate::kernel_tests::tx::ExecutionOutputExt;
@@ -212,142 +204,109 @@ fn test_compute_output_note_id() -> anyhow::Result<()> {
     Ok(())
 }
 
-#[test]
-fn test_epilogue_asset_preservation_violation_too_few_input() -> anyhow::Result<()> {
+/// Tests that a transaction fails due to the asset preservation rules when the input note has an
+/// asset with amount 100 and the output note has the same asset with amount 200.
+#[tokio::test]
+async fn epilogue_fails_when_num_output_assets_exceed_num_input_assets() -> anyhow::Result<()> {
+    // Create an input asset with amount 100 and an output asset with amount 200.
+    let input_asset = FungibleAsset::new(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1.try_into()?, 100)?;
+    let output_asset = input_asset.add(input_asset)?;
+
     let mut builder = MockChain::builder();
-    let account = builder
-        .add_existing_mock_account_with_assets(Auth::IncrNonce, AssetVault::mock().assets())?;
+    let account = builder.add_existing_mock_account(Auth::IncrNonce)?;
+    // Add an input note that (automatically) adds its assets to the transaction's input vault, but
+    // _does not_ add the asset to the account. This is just to keep the test conceptually simple -
+    // there is no account involved.
+    let input_note = NoteBuilder::new(account.id(), *builder.rng_mut())
+        .add_assets([Asset::from(input_asset)])
+        .build()?;
+    builder.add_output_note(OutputNote::Full(input_note.clone()));
     let mock_chain = builder.build()?;
-
-    let fungible_asset_1: Asset = FungibleAsset::new(
-        ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1.try_into()?,
-        CONSUMED_ASSET_1_AMOUNT,
-    )?
-    .into();
-    let fungible_asset_2: Asset = FungibleAsset::new(
-        ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_2.try_into()?,
-        CONSUMED_ASSET_2_AMOUNT,
-    )?
-    .into();
-
-    let output_note_1 = NoteBuilder::new(account.id(), rng())
-        .add_assets([fungible_asset_1])
-        .dynamically_linked_libraries(TransactionKernel::mock_libraries())
-        .build()?;
-    let output_note_2 = NoteBuilder::new(account.id(), rng())
-        .add_assets([fungible_asset_2])
-        .dynamically_linked_libraries(TransactionKernel::mock_libraries())
-        .build()?;
-
-    let input_note = create_spawn_note([&output_note_1, &output_note_2])?;
-
-    let tx_context = mock_chain
-        .build_tx_context(TxContextInput::AccountId(account.id()), &[], &[input_note])?
-        .extend_expected_output_notes(vec![
-            OutputNote::Full(output_note_1),
-            OutputNote::Full(output_note_2),
-        ])
-        .build()?;
-
-    let output_notes_data_procedure =
-        create_mock_notes_procedure(tx_context.expected_output_notes());
 
     let code = format!(
         "
-        use.$kernel::prologue
-        use.mock::account
-        use.$kernel::epilogue
+      use.mock::account
+      use.mock::util
 
-        {output_notes_data_procedure}
-
-        begin
-            exec.prologue::prepare_transaction
-            exec.create_mock_notes
-            exec.epilogue::finalize_transaction
-
-            # truncate the stack
-            movupw.3 dropw movupw.3 dropw movup.9 drop
-        end
-        "
+      begin
+          # create a note with the output asset
+          push.{OUTPUT_ASSET}
+          exec.util::create_random_note_with_asset drop
+          # => []
+      end
+      ",
+        OUTPUT_ASSET = Word::from(output_asset),
     );
 
-    let exec_output = tx_context.execute_code_blocking(&code);
+    let builder = ScriptBuilder::with_mock_libraries()?;
+    let source_manager = builder.source_manager();
+    let tx_script = builder.compile_tx_script(code)?;
 
-    assert_execution_error!(exec_output, ERR_EPILOGUE_TOTAL_NUMBER_OF_ASSETS_MUST_STAY_THE_SAME);
+    let tx_context = mock_chain
+        .build_tx_context(TxContextInput::AccountId(account.id()), &[], &[input_note])?
+        .tx_script(tx_script)
+        .with_source_manager(source_manager)
+        .build()?;
+
+    let exec_output = tx_context.execute().await;
+    assert_transaction_executor_error!(
+        exec_output,
+        ERR_EPILOGUE_TOTAL_NUMBER_OF_ASSETS_MUST_STAY_THE_SAME
+    );
+
     Ok(())
 }
 
-#[test]
-fn test_epilogue_asset_preservation_violation_too_many_fungible_input() -> anyhow::Result<()> {
+/// Tests that a transaction fails due to the asset preservation rules when the input note has an
+/// asset with amount 200 and the output note has the same asset with amount 100.
+#[tokio::test]
+async fn epilogue_fails_when_num_input_assets_exceed_num_output_assets() -> anyhow::Result<()> {
+    // Create an input asset with amount 200 and an output asset with amount 100.
+    let output_asset = FungibleAsset::new(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1.try_into()?, 100)?;
+    let input_asset = output_asset.add(output_asset)?;
+
     let mut builder = MockChain::builder();
-    let account = builder
-        .add_existing_mock_account_with_assets(Auth::IncrNonce, AssetVault::mock().assets())?;
+    let account = builder.add_existing_mock_account(Auth::IncrNonce)?;
+    // Add an input note that (automatically) adds its assets to the transaction's input vault, but
+    // _does not_ add the asset to the account. This is just to keep the test conceptually simple -
+    // there is no account involved.
+    let input_note = NoteBuilder::new(account.id(), *builder.rng_mut())
+        .add_assets([Asset::from(output_asset)])
+        .build()?;
+    builder.add_output_note(OutputNote::Full(input_note.clone()));
     let mock_chain = builder.build()?;
-
-    let fungible_asset_1: Asset = FungibleAsset::new(
-        ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1.try_into()?,
-        CONSUMED_ASSET_1_AMOUNT,
-    )?
-    .into();
-    let fungible_asset_2: Asset = FungibleAsset::new(
-        ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_2.try_into()?,
-        CONSUMED_ASSET_2_AMOUNT,
-    )?
-    .into();
-    let fungible_asset_3: Asset = FungibleAsset::new(
-        ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_3.try_into()?,
-        CONSUMED_ASSET_3_AMOUNT,
-    )?
-    .into();
-
-    let output_note_1 = NoteBuilder::new(account.id(), rng())
-        .add_assets([fungible_asset_1])
-        .dynamically_linked_libraries(TransactionKernel::mock_libraries())
-        .build()?;
-    let output_note_2 = NoteBuilder::new(account.id(), rng())
-        .add_assets([fungible_asset_2])
-        .dynamically_linked_libraries(TransactionKernel::mock_libraries())
-        .build()?;
-    let output_note_3 = NoteBuilder::new(account.id(), rng())
-        .add_assets([fungible_asset_3])
-        .dynamically_linked_libraries(TransactionKernel::mock_libraries())
-        .build()?;
-
-    let input_note = create_spawn_note([&output_note_1, &output_note_2, &output_note_3])?;
-
-    let tx_context = mock_chain
-        .build_tx_context(TxContextInput::AccountId(account.id()), &[], &[input_note])?
-        .extend_expected_output_notes(vec![
-            OutputNote::Full(output_note_1),
-            OutputNote::Full(output_note_2),
-        ])
-        .build()?;
-
-    let output_notes_data_procedure =
-        create_mock_notes_procedure(tx_context.expected_output_notes());
 
     let code = format!(
         "
-        use.$kernel::prologue
-        use.mock::account
-        use.$kernel::epilogue
+      use.mock::account
+      use.mock::util
 
-        {output_notes_data_procedure}
-
-        begin
-            exec.prologue::prepare_transaction
-            exec.create_mock_notes
-            exec.epilogue::finalize_transaction
-
-            # truncate the stack
-            movupw.3 dropw movupw.3 dropw movup.9 drop
-        end
-        "
+      begin
+          # create a note with the output asset
+          push.{OUTPUT_ASSET}
+          exec.util::create_random_note_with_asset drop
+          # => []
+      end
+      ",
+        OUTPUT_ASSET = Word::from(input_asset),
     );
 
-    let exec_output = tx_context.execute_code_blocking(&code);
+    let builder = ScriptBuilder::with_mock_libraries()?;
+    let source_manager = builder.source_manager();
+    let tx_script = builder.compile_tx_script(code)?;
 
-    assert_execution_error!(exec_output, ERR_EPILOGUE_TOTAL_NUMBER_OF_ASSETS_MUST_STAY_THE_SAME);
+    let tx_context = mock_chain
+        .build_tx_context(TxContextInput::AccountId(account.id()), &[], &[input_note])?
+        .tx_script(tx_script)
+        .with_source_manager(source_manager)
+        .build()?;
+
+    let exec_output = tx_context.execute().await;
+    assert_transaction_executor_error!(
+        exec_output,
+        ERR_EPILOGUE_TOTAL_NUMBER_OF_ASSETS_MUST_STAY_THE_SAME
+    );
+
     Ok(())
 }