+++ crates/miden-testing/src/kernel_tests/tx/test_note.rs
@@ -1,6 +1,5 @@
 use alloc::collections::BTreeMap;
 use alloc::sync::Arc;
-use alloc::vec::Vec;
 
 use anyhow::Context;
 use miden_lib::account::wallets::BasicWallet;
@@ -30,7 +29,7 @@ use miden_objects::testing::account_id::{
     ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE,
     ACCOUNT_ID_SENDER,
 };
-use miden_objects::transaction::{AccountInputs, OutputNote, TransactionArgs};
+use miden_objects::transaction::{OutputNote, TransactionArgs};
 use miden_objects::{Felt, Word, ZERO};
 use miden_processor::fast::ExecutionOutput;
 use rand::SeedableRng;
@@ -155,11 +154,8 @@ fn test_note_script_and_note_args() -> miette::Result<()> {
         (tx_context.input_notes().get_note(1).note().id(), note_args[0]),
     ]);
 
-    let tx_args = TransactionArgs::new(
-        tx_context.tx_args().advice_inputs().clone().map,
-        Vec::<AccountInputs>::new(),
-    )
-    .with_note_args(note_args_map);
+    let tx_args = TransactionArgs::new(tx_context.tx_args().advice_inputs().clone().map)
+        .with_note_args(note_args_map);
 
     tx_context.set_tx_args(tx_args);
     let exec_output = tx_context.execute_code_blocking(code).unwrap();