+++ crates/miden-testing/src/mock_host.rs
@@ -63,6 +63,9 @@ impl<'store> MockHost<'store> {
                 &TransactionEvent::AccountPushProcedureIndex,
                 &TransactionEvent::LinkMapSet,
                 &TransactionEvent::LinkMapGet,
+                // TODO: It should be possible to remove this after implementing
+                // https://github.com/0xMiden/miden-base/issues/1852.
+                &TransactionEvent::EpilogueBeforeTxFeeRemovedFromAccount,
             ]
             .map(TransactionEvent::event_id),
         );
@@ -78,6 +81,7 @@ impl<'store> MockHost<'store> {
                 &TransactionEvent::AccountVaultBeforeGetBalance,
                 &TransactionEvent::AccountVaultBeforeHasNonFungibleAsset,
                 &TransactionEvent::AccountVaultBeforeAddAsset,
+                &TransactionEvent::AccountVaultBeforeRemoveAsset,
                 &TransactionEvent::AccountStorageBeforeSetMapItem,
                 &TransactionEvent::AccountStorageBeforeGetMapItem,
             ]