+++ crates/miden-testing/src/tx_context/builder.rs
@@ -9,31 +9,19 @@ use anyhow::Context;
 use miden_lib::testing::account_component::IncrNonceAuthComponent;
 use miden_lib::testing::mock_account::MockAccountExt;
 use miden_objects::EMPTY_WORD;
-use miden_objects::account::{
-    Account,
-    AccountHeader,
-    AccountId,
-    PartialAccount,
-    PartialStorage,
-    PartialStorageMap,
-    PublicKeyCommitment,
-    Signature,
-    StorageSlot,
-};
+use miden_objects::account::{Account, AccountHeader, AccountId, PublicKeyCommitment, Signature};
 use miden_objects::assembly::DefaultSourceManager;
 use miden_objects::assembly::debuginfo::SourceManagerSync;
-use miden_objects::asset::PartialVault;
+use miden_objects::block::AccountWitness;
 use miden_objects::note::{Note, NoteId, NoteScript};
 use miden_objects::testing::account_id::ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE;
 use miden_objects::testing::noop_auth_component::NoopAuthComponent;
 use miden_objects::transaction::{
-    AccountInputs,
     OutputNote,
     TransactionArgs,
     TransactionInputs,
     TransactionScript,
 };
-use miden_objects::vm::AdviceMap;
 use miden_processor::{AdviceInputs, Felt, Word};
 use miden_tx::TransactionMastStore;
 use miden_tx::auth::BasicAuthenticator;
@@ -86,7 +74,7 @@ pub struct TransactionContextBuilder {
     advice_inputs: AdviceInputs,
     authenticator: Option<MockAuthenticator>,
     expected_output_notes: Vec<Note>,
-    foreign_account_inputs: BTreeMap<AccountId, AccountInputs>,
+    foreign_account_inputs: BTreeMap<AccountId, (Account, AccountWitness)>,
     input_notes: Vec<Note>,
     tx_script: Option<TransactionScript>,
     tx_script_args: Word,
@@ -114,7 +102,7 @@ impl TransactionContextBuilder {
             foreign_account_inputs: BTreeMap::new(),
             auth_args: EMPTY_WORD,
             signatures: Vec::new(),
-            is_lazy_loading_enabled: false,
+            is_lazy_loading_enabled: true,
             note_scripts: BTreeMap::new(),
         }
     }
@@ -177,9 +165,13 @@ impl TransactionContextBuilder {
     }
 
     /// Set foreign account codes that are used by the transaction
-    pub fn foreign_accounts(mut self, inputs: impl IntoIterator<Item = AccountInputs>) -> Self {
-        self.foreign_account_inputs
-            .extend(inputs.into_iter().map(|account_inputs| (account_inputs.id(), account_inputs)));
+    pub fn foreign_accounts(
+        mut self,
+        inputs: impl IntoIterator<Item = (Account, AccountWitness)>,
+    ) -> Self {
+        self.foreign_account_inputs.extend(
+            inputs.into_iter().map(|(account, witness)| (account.id(), (account, witness))),
+        );
         self
     }
 
@@ -218,20 +210,10 @@ impl TransactionContextBuilder {
         self
     }
 
-    /// Causes the transaction to only construct a minimal partial account as the transaction
-    /// input, causing lazy loading of assets and storage map items throughout transaction
-    /// execution. Additionally, foreign accounts aren't provided via the transaction args but are
-    /// lazy loaded as well.
-    ///
-    /// This exists to test lazy loading selectively and should go away in the future.
-    pub fn enable_lazy_loading(mut self) -> Self {
-        self.is_lazy_loading_enabled = true;
-        self
-    }
-
     /// Disables lazy loading.
     ///
-    /// This is the opposite of [`Self::enable_lazy_loading`] - see its docs for details.
+    /// Only affects [`TransactionContext::execute_code`] and causes the host to _not_ handle lazy
+    /// loading events.
     pub fn disable_lazy_loading(mut self) -> Self {
         self.is_lazy_loading_enabled = false;
         self
@@ -286,7 +268,7 @@ impl TransactionContextBuilder {
     /// If no transaction inputs were provided manually, an ad-hoc MockChain is created in order
     /// to generate valid block data for the required notes.
     pub fn build(self) -> anyhow::Result<TransactionContext> {
-        let tx_inputs = match self.tx_inputs {
+        let mut tx_inputs = match self.tx_inputs {
             Some(tx_inputs) => tx_inputs,
             None => {
                 // If no specific transaction inputs was provided, initialize an ad-hoc mockchain
@@ -310,14 +292,8 @@ impl TransactionContextBuilder {
             },
         };
 
-        let foreign_account_inputs = if self.is_lazy_loading_enabled {
-            Vec::new()
-        } else {
-            self.foreign_account_inputs.values().cloned().collect()
-        };
+        let tx_args = TransactionArgs::default().with_note_args(self.note_args);
 
-        let tx_args = TransactionArgs::new(AdviceMap::default(), foreign_account_inputs)
-            .with_note_args(self.note_args);
         let mut tx_args = if let Some(tx_script) = self.tx_script {
             tx_args.with_tx_script_and_args(tx_script, self.tx_script_args)
         } else {
@@ -331,26 +307,14 @@ impl TransactionContextBuilder {
             tx_args.add_signature(public_key_commitment, message, signature);
         }
 
-        // If partial loading is enabled, construct an account that doesn't contain all
-        // merkle paths of assets and storage maps, in order to test lazy loading.
-        // Otherwise, load the full account.
-        let tx_inputs = if self.is_lazy_loading_enabled {
-            let (_, block_header, partial_blockchain, input_notes, _) = tx_inputs.into_parts();
-            // Note that we use self.account instead of account, because we cannot do the same
-            // operation on a partial vault.
-            let account = minimal_partial_account(&self.account)?;
-            TransactionInputs::new(account, block_header, partial_blockchain, input_notes)?
-                .with_tx_args(tx_args)
-        } else {
-            tx_inputs.with_tx_args(tx_args)
-        };
+        tx_inputs.set_tx_args(tx_args);
 
         let mast_store = {
             let mast_forest_store = TransactionMastStore::new();
             mast_forest_store.load_account_code(tx_inputs.account().code());
 
-            for acc_inputs in self.foreign_account_inputs.values() {
-                mast_forest_store.insert(acc_inputs.code().mast());
+            for (account, _) in self.foreign_account_inputs.values() {
+                mast_forest_store.insert(account.code().mast());
             }
 
             mast_forest_store
@@ -374,44 +338,4 @@ impl Default for TransactionContextBuilder {
     fn default() -> Self {
         Self::with_existing_mock_account()
     }
-}
-
-/// Creates a minimal [`PartialAccount`] from the provided full [`Account`].
-fn minimal_partial_account(account: &Account) -> anyhow::Result<PartialAccount> {
-    // Construct a partial vault that tracks the empty word, but none of the assets
-    // that are actually in the asset tree. That way, the partial vault has the same
-    // root as the full vault, but will not add any relevant merkle paths to the
-    // merkle store, which will test lazy loading of assets.
-    let mut partial_vault = PartialVault::default();
-    partial_vault.add(account.vault().open(Word::empty()))?;
-
-    // Construct a partial storage that tracks the empty word in all storage maps, but none
-    // of the other keys, following the same rationale as the partial vault above.
-    let storage_header = account.storage().to_header();
-    let storage_maps =
-        account.storage().slots().iter().filter_map(|storage_slot| match storage_slot {
-            StorageSlot::Map(storage_map) => {
-                let mut partial_storage_map = PartialStorageMap::default();
-                let key = Word::empty();
-                let witness = storage_map.open(&key);
-                partial_storage_map
-                    .add(witness)
-                    .expect("adding the first proof should never error");
-                Some(partial_storage_map)
-            },
-            _ => None,
-        });
-    let partial_storage = PartialStorage::new(storage_header, storage_maps)
-        .expect("provided storage maps should match storage header");
-
-    let account = PartialAccount::new(
-        account.id(),
-        account.nonce(),
-        account.code().clone(),
-        partial_storage,
-        partial_vault,
-        None,
-    )?;
-
-    Ok(account)
 }